+++ playback/src/decoder/libvorbis_decoder.rs
@@ -38,14 +38,11 @@ where
         loop {
             match self.0.packets().next() {
                 Some(Ok(packet)) => {
-                    // Losslessly represent [-32768, 32767] to [-1.0, 1.0] while maintaining DC linearity.
                     return Ok(Some(AudioPacket::Samples(
                         packet
                             .data
                             .iter()
-                            .map(|sample| {
-                                ((*sample as f64 + 0.5) / (std::i16::MAX as f64 + 0.5)) as f32
-                            })
+                            .map(|sample| (*sample as f64 / 0x8000 as f64) as f32)
                             .collect(),
                     )));
                 }